refactor: extract preview fakes to dedicated login.preview package
